chore: Reorganize Features/Brands/ and Features/Dashboards/ folders
