+++ BlazorHero.CleanArchitecture.Application/Features/Brands/Queries/GetById/GetBrandByIdResponse.cs
