+++ BlazorHero.CleanArchitecture.Application/Mappings/BrandProfile.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using BlazorHero.CleanArchitecture.Application.Features.Brands.AddEdit;
+using BlazorHero.CleanArchitecture.Application.Features.Brands.Commands.AddEdit;
 using BlazorHero.CleanArchitecture.Application.Features.Brands.Queries.GetAll;
 using BlazorHero.CleanArchitecture.Application.Features.Brands.Queries.GetById;
 using BlazorHero.CleanArchitecture.Domain.Entities.Catalog;
