+++ BlazorHero.CleanArchitecture.Application/Resources/Features/Brands/Commands/Delete/DeleteBrandCommandHandler.en.resx
