+++ BlazorHero.CleanArchitecture.Application/Resources/Features/Brands/Commands/Delete/DeleteBrandCommandHandler.it.resx
