+++ BlazorHero.CleanArchitecture.Application/Resources/Features/Brands/Commands/Delete/DeleteBrandCommandHandler.ru.resx
