+++ BlazorHero.CleanArchitecture.Application/Resources/Features/Dashboards/Queries/GetData/GetDashboardDataQueryHandler.en.resx
