+++ BlazorHero.CleanArchitecture.Application/Resources/Validators/Features/Brands/Commands/AddEdit/AddEditBrandCommandValidator.fr.resx
