+++ BlazorHero.CleanArchitecture.Client.Infrastructure/Managers/Dashboard/DashboardManager.cs
@@ -1,8 +1,8 @@
-﻿using BlazorHero.CleanArchitecture.Application.Features.Dashboard.GetData;
-using BlazorHero.CleanArchitecture.Client.Infrastructure.Extensions;
+﻿using BlazorHero.CleanArchitecture.Client.Infrastructure.Extensions;
 using BlazorHero.CleanArchitecture.Shared.Wrapper;
 using System.Net.Http;
 using System.Threading.Tasks;
+using BlazorHero.CleanArchitecture.Application.Features.Dashboards.Queries.GetData;
 
 namespace BlazorHero.CleanArchitecture.Client.Infrastructure.Managers.Dashboard
 {
