+++ BlazorHero.CleanArchitecture.Client.Infrastructure/Managers/Dashboard/IDashboardManager.cs
@@ -1,6 +1,6 @@
-﻿using BlazorHero.CleanArchitecture.Application.Features.Dashboard.GetData;
-using BlazorHero.CleanArchitecture.Shared.Wrapper;
+﻿using BlazorHero.CleanArchitecture.Shared.Wrapper;
 using System.Threading.Tasks;
+using BlazorHero.CleanArchitecture.Application.Features.Dashboards.Queries.GetData;
 
 namespace BlazorHero.CleanArchitecture.Client.Infrastructure.Managers.Dashboard
 {
