+++ BlazorHero.CleanArchitecture/Client/Pages/Catalog/AddEditBrandModal.razor
@@ -1,4 +1,4 @@
-﻿@using BlazorHero.CleanArchitecture.Application.Features.Brands.AddEdit
+﻿@using BlazorHero.CleanArchitecture.Application.Features.Brands.Commands.AddEdit
 <EditForm Model="@AddEditBrandModel" OnValidSubmit="SaveAsync">
     <FluentValidationValidator @ref="_fluentValidationValidator" />
     <MudDialog>
